feat: Move qr code to details screen
Changes based on the updated design in Figma:

- The qr code was previously shown when clicking a "show for inspection"
  button. This is instead moved to the ticket details screen.
- THe ticket details was previously navigated to through a click on the
  three-dottet button in the upper right corner of the ticket in the
  ticket list. This button was now removed, and instead a section link
  item is used for navigating to the ticket details.
