+++ src/screens/Ticketing/Ticket/index.tsx
@@ -12,14 +12,12 @@ type Props = {
   fareContract: FareContract;
   now: number;
   onPressDetails: () => void;
-  onPressInspection: () => void;
 };
 
 const SimpleTicket: React.FC<Props> = ({
   fareContract: fc,
   now,
   onPressDetails,
-  onPressInspection,
 }) => {
   const nowSeconds = now / 1000;
   const isNotExpired = fc.usage_valid_to >= nowSeconds;
@@ -36,7 +34,6 @@ const SimpleTicket: React.FC<Props> = ({
           validTo={fc.usage_valid_to}
           isNotExpired={isNotExpired}
           isRefunded={isRefunded}
-          onPressDetails={onPressDetails}
         />
         {isValidTicket ? (
           <ValidityLine
@@ -51,13 +48,14 @@ const SimpleTicket: React.FC<Props> = ({
 
         <TicketInfo fareContract={fc} />
       </Sections.GenericItem>
-      {isValidTicket && (
-        <Sections.LinkItem
-          text={t(TicketTexts.controlLink)}
-          onPress={onPressInspection}
-          disabled={!fc.qr_code}
-        />
-      )}
+      <Sections.LinkItem
+        text={t(
+          isValidTicket
+            ? TicketTexts.detailsLink.valid
+            : TicketTexts.detailsLink.expired,
+        )}
+        onPress={onPressDetails}
+      />
     </Sections.Section>
   );
 };
